requirements.txt
# python>=3.6
Django>=2.1
drf-yasg>=1.15.0
djangorestframework>=3.9.1
redis>=2.19.1
pymysql>=0.7.11
gevent>=1.2.1

# admin皮肤
django-simpleui
